+++ src/dpr/components/_charts/utils.ts
@@ -9,7 +9,6 @@ import {
   ChartDetails,
   ChartMetaData,
 } from '../../types/Charts'
-import { createTimeseriesMatrixChart } from './chart/heatmap/utils'
 import { DashboardDataResponse } from '../../types/Metrics'
 import {
   BarChartVisualisationColumn,
@@ -19,6 +18,7 @@ import {
 import DatasetHelper from '../../utils/datasetHelper'
 import DashboardListUtils from '../_dashboards/dashboard-list/utils'
 import { Granularity } from '../_inputs/granular-date-range/types'
+import HeatmapChart from './chart/heatmap/Heatmap'
 
 dayjs.extend(weekOfYear)
 
@@ -83,7 +83,7 @@ export const createMatrixChart = (
 
   const { latestData, dataSetRows, timeseriesData } = getDataForTimeseriesCharts(chartDefinition, rawData)
   if (dataSetRows.length) {
-    chart = createTimeseriesMatrixChart(chartDefinition, timeseriesData, granularity)
+    chart = new HeatmapChart(timeseriesData, granularity, chartDefinition).build()
     table = createTimeseriesTable(chartDefinition, timeseriesData)
     details = getChartDetails(chartDefinition, latestData, true)
   }