+++ src/dpr/components/_dashboards/dashboard-list/utils.ts
@@ -14,8 +14,9 @@ export const createList = (
   dashboardData: DashboardDataResponse[],
 ): { table: MoJTable; ts: string } => {
   const { columns, options } = listDefinition
-  const showLatest = (<ListDashboardVisualisationOptions>options)?.showLatest || true
-  const columnsAsList = (<ListDashboardVisualisationOptions>options)?.columnsAsList
+  const listOptions = <ListDashboardVisualisationOptions>options
+  const showLatest = listOptions?.showLatest !== undefined ? listOptions.showLatest : true
+  const columnsAsList = listOptions?.columnsAsList
   const { measures, keys } = columns
   const showAllData = (!measures && !keys) || (!measures.length && !keys)